Changed all float to double (vec3f_t to vec3d_t) due to float inaccuracies.

+++ src/engine/engine.cpp
@@ -33,7 +33,7 @@ void Engine::start()
 	realTime = realDate;
 }
 
-void Engine::update(float dt)
+void Engine::update(double dt)
 {
 	// Update current julian time
 	realTime += dt / SECONDS_PER_DAY;
